Revert "cleaning incompatibilities"
This reverts commit e4a4817ff2b569ee5462897d53ea12fe74010dbf.

+++ grblCA/settings.c
@@ -353,7 +353,7 @@ return ((1 << A_DIRECTION_BIT));
 }
 
 // Returns limit pin mask according to Grbl internal axis indexing.
-uint16_t get_limit_pin_mask(uint8_t axis_idx)
+uint8_t get_limit_pin_mask(uint8_t axis_idx)
 {
   if (axis_idx == X_AXIS)
   {
